+++ packages/errors-nest/src/errors/common.errors.ts
@@ -1,4 +1,7 @@
 import { BadRequestError } from '@volontariapp/errors';
 
-export const INVALID_DATE_PARAMETERS = (details: string) =>
-  new BadRequestError(`Invalid date parameter: ${details}`, 'INVALID_DATE_PARAMETER');
+export const INVALID_DATE_PARAMETERS = (details?: string) =>
+  new BadRequestError(
+    `Invalid date parameter${details !== undefined && details !== '' ? `: ${details}` : ''}`,
+    'INVALID_DATE_PARAMETER',
+  );