+++ packages/validation-nest/src/interfaces/grpc-validation-options.interface.ts
@@ -1,0 +1,5 @@
+import type { ValidationPipeOptions } from '@nestjs/common';
+
+export interface GrpcValidationOptions extends ValidationPipeOptions {
+  enumMaps?: Record<string, Record<string | number, string | number>>;
+}