+++ packages/domain-event/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 2.2.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors@0.4.3
+  - @volontariapp/errors-nest@0.7.3
+
+## 2.2.7
+
+### Patch Changes
+
+- including google GRPC timestamps + update mappers
+- Updated dependencies
+  - @volontariapp/errors-nest@0.7.2
+
 ## 2.2.6
 
 ### Patch Changes