+++ packages/errors-nest/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.7.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors@0.4.3
+
+## 0.7.2
+
+### Patch Changes
+
+- including google GRPC timestamps + update mappers
+
 ## 0.7.1
 
 ### Patch Changes