+++ packages/errors/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.4.3
+
+### Patch Changes
+
+- adding new database errors
+
 ## 0.4.2
 
 ### Patch Changes