+++ packages/health-check-nest/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 0.1.13
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors-nest@0.7.3
+  - @volontariapp/health-check@0.1.5
+
+## 0.1.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @volontariapp/errors-nest@0.7.2
+
 ## 0.1.11
 
 ### Patch Changes